+++ src/main/java/app/repository/Jogador1Repository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.entity.Jogador1;
 
-public interface Jogador1Repository extends JpaRepository<Jogador1, Long> {
+public interface Jogador1Repository extends JpaRepository<Jogador1, Integer> {
 	
 }
