+++ src/main/java/app/service/RankingService.java
@@ -19,7 +19,7 @@ public class RankingService {
     }
     
     
-    public String update(long idPontuacao, Ranking ranking) {
+    public String update(Integer idPontuacao, Ranking ranking) {
         ranking.setIdPontuacao(idPontuacao);
         this.rankingRepository.save(ranking);        
         return "Atualizado com sucesso!";
@@ -31,13 +31,13 @@ public class RankingService {
     }
     
     
-    public Ranking findById(long idPontuacao) {
+    public Ranking findById(Integer idPontuacao) {
         Ranking ranking = this.rankingRepository.findById(idPontuacao).get();
         return ranking;
     }
     
         
-    public String delete(long idPontuacao) {
+    public String delete(Integer idPontuacao) {
         this.rankingRepository.deleteById(idPontuacao);        
         return "Deletado com sucesso!";
     }
